Organizing components in respective files as well as cleaning up styles.
